app/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2